implement column features

+++ src/HealthApp.Api/Commons/PagingResult.cs
@@ -3,13 +3,13 @@
 public class PagingResult<T>
 {
     public IReadOnlyList<T> Items { get; init; } = [];
-    public long TotalCount { get; init; }
+    public int TotalCount { get; init; }
     public int Page { get; init; }
     public int PageSize { get; init; }
 
     public PagingResult() { }
 
-    public PagingResult(IReadOnlyList<T> items, long totalCount, int page, int pageSize)
+    public PagingResult(IReadOnlyList<T> items, int totalCount, int page, int pageSize)
     {
         Items = items;
         TotalCount = totalCount;
@@ -17,6 +17,6 @@ public class PagingResult<T>
         PageSize = pageSize;
     }
 
-    public static PagingResult<T> Create(IReadOnlyList<T> items, long totalCount, int page, int pageSize)
+    public static PagingResult<T> Create(IReadOnlyList<T> items, int totalCount, int page, int pageSize)
         => new(items, totalCount, page, pageSize);
 }
