+++ src/HealthApp.Application/Features/Exercises/IExerciseService.cs
@@ -3,7 +3,7 @@
 public interface IExerciseService
 {
     Task<CreateExerciseResponse> CreateAsync(CreateExerciseRequest req, CancellationToken ct = default);
-    Task<(IReadOnlyList<ExerciseSummaryDto> Items, long TotalCount)> FetchByProfileAsync(long profileId, DateOnly? byDate, int page, int pageSize, CancellationToken ct = default);
+    Task<(IReadOnlyList<ExerciseSummaryDto> Items, int TotalCount)> FetchByProfileAsync(long profileId, DateOnly? byDate, int page, int pageSize, CancellationToken ct = default);
     Task<ExerciseSummaryDto?> UpdateAsync(long id, UpdateExerciseRequest req, CancellationToken ct = default);
     Task<bool> DeleteAsync(long id, CancellationToken ct = default);
     Task<ExerciseAchievementDto> GetAchievementByDateAsync(long profileId, DateOnly day, CancellationToken ct = default);
