+++ src/HealthApp.Application/Features/Meals/IMealService.cs
@@ -3,7 +3,7 @@
 public interface IMealService
 {
     Task<CreateMealResponse> CreateAsync(CreateMealRequest req, CancellationToken ct = default);
-    Task<(IReadOnlyList<MealSummaryDto> Items, long TotalCount)> FetchByProfileAsync(
+    Task<(IReadOnlyList<MealSummaryDto> Items, int TotalCount)> FetchByProfileAsync(
         long profileId,
         EnumMealType? type,
         int page,
