+++ src/HealthApp.Application/Features/Profiles/IProfileService.cs
@@ -3,7 +3,7 @@
 public interface IProfileService
 {
     public Task<CreateProfileResponse> CreateProfileAsync(CreateProfileRequest req, CancellationToken ct = default);
-    Task<(IReadOnlyList<ProfileSummaryDto> Items, long TotalCount)> FetchProfilesByUserAsync(
+    Task<(IReadOnlyList<ProfileSummaryDto> Items, int TotalCount)> FetchProfilesByUserAsync(
         long userId,
         int page,
         int pageSize,
