+++ src/HealthApp.Domain/Models/ExerciseModels/IExerciseRepository.cs
@@ -6,7 +6,7 @@ public interface IExerciseRepository : IRepository<Exercise>
     Task<Exercise?> UpdateAsync(ExerciseSummaryDto dto, CancellationToken ct = default);
     Task<Exercise?> FindByIdAsync(long id, CancellationToken ct = default);
     Task<bool> DeleteAsync(long id, CancellationToken ct = default);
-    Task<(IReadOnlyList<TResult> Items, long TotalCount)> FetchByProfileIdAsync<TResult>(
+    Task<(IReadOnlyList<TResult> Items, int TotalCount)> FetchByProfileIdAsync<TResult>(
         ISimplePagingSpecification<Exercise, TResult> spec,
         CancellationToken ct = default
     );
