+++ src/HealthApp.Domain/Models/MealModels/IMealRepository.cs
@@ -6,7 +6,7 @@ public interface IMealRepository : IRepository<Meal>
     Task<Meal?> UpdateAsync(MealSummaryDto dto, CancellationToken ct = default);
     Task<Meal?> FindByIdAsync(long id, CancellationToken ct = default);
     Task<bool> DeleteAsync(long id, CancellationToken ct = default);
-    Task<(IReadOnlyList<TResult> Items, long TotalCount)> FetchByProfileIdAsync<TResult>(
+    Task<(IReadOnlyList<TResult> Items, int TotalCount)> FetchByProfileIdAsync<TResult>(
         ISimplePagingSpecification<Meal, TResult> spec,
         CancellationToken ct = default
     );
